added file uploads as possible form field
